Fix eslint errors
