+++ src/logic/operate.js
@@ -1,14 +1,22 @@
-import operate from './operate';
+import Big from 'big.js';
 
-const calculate = (data, buttonName) => {
-  const copyData = { ...data };
-  if (buttonName === '+/-') {
-    copyData.total *= -1;
-    copyData.next *= -1;
-  } else if (buttonName === '+' || buttonName === '-' || buttonName === 'x' || buttonName === '÷' || buttonName === '%') {
-    copyData.total = operate(copyData.total, copyData.next, buttonName);
+const operate = (numberOne, numberTwo, operation) => {
+  if (operation === '+') {
+    return Big(numberOne) * 1 + Big(numberTwo) * 1;
   }
-  return copyData;
+  if (operation === '-') {
+    return Big(numberOne) - Big(numberTwo);
+  }
+  if (operation === 'x') {
+    return Big(numberOne) * Big(numberTwo);
+  }
+  if (operation === '÷') {
+    return Big(numberOne) / Big(numberTwo);
+  }
+  if (operation === '%') {
+    return Big(numberOne) * 0.01;
+  }
+  return '0';
 };
 
-export default calculate;
+export default operate;
